pyproject.toml
[project]
name = "web-subscription"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi[standard]>=0.118.0",
    "google-auth>=2.41.1",
    "google-cloud-firestore>=2.21.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "stripe>=13.0.0",
    "uvicorn>=0.37.0",
]
